requirements.txt
faster-whisper>=1.0.0
groq>=0.4.0
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
python-dotenv>=1.0.0
rumps>=0.4.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-AVFoundation>=9.0
